# packages/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/stations": {
            "get": {
                "description": "Retrieves information about all available stations.\nThis endpoint returns a list of all stations and their details.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "Get all stations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/utils.Station"
                            }
                        }
                    },
                    "304": {
                        "description": "Not Modified: The ETag matches the If-None-Match header."
                    },
                    "404": {
                        "description": "Not Found: The specified network does not exist.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Error retrieving stations data."
                    }
                }
            }
        },
        "/stations/search": {
            "get": {
                "description": "Searches for a station using the provided name and returns the matching station information.\nThis endpoint is case and whitespace insensitive. Station names are not unique, so when several stations carry the name the one with the lowest id is returned.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "Search for a station by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the station to search for",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully found and retrieved the station data.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/utils.StationListEntry"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request: Missing station name parameter.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Station not found: No station matches the provided name.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stations/{stationId}": {
            "get": {
                "description": "Retrieves information about a specific station based on its ID.\nThis endpoint returns the details of a single station.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "Get a single station by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the station",
                        "name": "stationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the specified station data.",
                        "schema": {
                            "$ref": "#/definitions/utils.StationListEntry"
                        }
                    },
                    "404": {
                        "description": "Station not found: The specified station does not exist.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v0/risk-prediction/segment-colors": {
            "get": {
                "description": "Retrieves risk predictions for transit segments.\nThis endpoint returns color-coded risk levels for different segments of the transit network based on recent ticket inspector activity.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prediction"
                ],
                "summary": "Get risk segments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved risk segments data",
                        "schema": {
                            "$ref": "#/definitions/v0.v0RiskData"
                        }
                    },
                    "404": {
                        "description": "Not Found: The specified network does not exist.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Failed to execute risk model"
                    }
                }
            }
        },
...
        "/v1/risk-prediction/segment-colors": {
            "get": {
                "description": "Retrieves risk predictions for transit segments.\nThis endpoint returns color-coded risk levels along with the risk value for different segments of the transit network based on recent ticket inspector activity.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prediction"
                ],
                "summary": "Get risk segments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved risk segments data",
                        "schema": {
                            "$ref": "#/definitions/prediction.RiskData"
                        }
                    },
                    "404": {
                        "description": "Not Found: The specified network does not exist.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Failed to execute risk model"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "prediction.RiskData": {
            "type": "object",
            "properties": {
                "segments_risk": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/prediction.SegmentRisk"
                    }
                }
            }
        },
        "prediction.SegmentRisk": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "risk": {
                    "type": "number"
                }
            }
        },
...
        "utils.Coordinates": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
        "utils.CoordinatesEntry": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
...
        "utils.Station": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "$ref": "#/definitions/utils.Coordinates"
                },
                "id": {
                    "type": "string"
                },
                "lines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "utils.StationListEntry": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "$ref": "#/definitions/utils.CoordinatesEntry"
                },
                "lines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "v0.v0RiskData": {
            "type": "object",
            "properties": {
                "last_modified": {
                    "type": "string"
                },
                "segment_colors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        }
    }
}